Move path file to map directory
